+++ CHANGELOG.md
@@ -6,7 +6,7 @@ this file.
 ### Changed
 * Updated documentation in the `README.md` file and the script
 ### Fixed
-* Fixed an error related to Hybrid MP4. This fixes the error: `258: bad argument #2 to 'gsub'`
+* Fixed an error related to Hybrid MP4. This fixes the error: `258: bad argument #2 to 'gsub'` (thanks to MATT_bauer, Hydraa, and vorngorth1 for reporting the crash related to this)
 
 ## 1.9 - 2024/05/03 14:45 GMT+08:00
 ### Added